+++ project/SudNFT/SudNFTQuickStart/src/main/java/tech/sud/mgp/hello/ui/nft/model/WalletInfoModel.java
@@ -9,7 +9,7 @@ import java.util.Objects;
  */
 public class WalletInfoModel implements Serializable {
 
-    public int type;
+    public long type;
     public String name;
     public String icon;
     public int zoneType;
